+++ src/components/manage/Tiles/Description/Edit.test.jsx
@@ -1,12 +1,28 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
+import configureStore from 'redux-mock-store';
+import { Provider } from 'react-intl-redux';
+
 import Edit from './Edit';
+
+const mockStore = configureStore();
 
 global.__SERVER__ = true; // eslint-disable-line no-underscore-dangle
 
 test('renders an edit description tile component', () => {
+  const store = mockStore({
+    intl: {
+      locale: 'en',
+      messages: {},
+    },
+  });
   const component = renderer.create(
-    <Edit properties={{ description: 'My Description' }} onChange={() => {}} />,
+    <Provider store={store}>
+      <Edit
+        properties={{ description: 'My Description' }}
+        onChange={() => {}}
+      />
+    </Provider>,
   );
   const json = component.toJSON();
   expect(json).toMatchSnapshot();