+++ src/components/manage/Tiles/Image/View.test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import View from './View';
+
+test('renders a view image component', () => {
+  const component = renderer.create(<View data={{ url: 'image.jpg' }} />);
+  const json = component.toJSON();
+  expect(json).toMatchSnapshot();
+});