+++ src/components/manage/Tiles/Title/Edit.jsx
@@ -8,6 +8,15 @@ import { Map } from 'immutable';
 import PropTypes from 'prop-types';
 import { stateFromHTML } from 'draft-js-import-html';
 import { Editor, DefaultDraftBlockRenderMap, EditorState } from 'draft-js';
+import { defineMessages, injectIntl, intlShape } from 'react-intl';
+import { Button, Icon } from 'semantic-ui-react';
+
+const messages = defineMessages({
+  title: {
+    id: 'Title',
+    defaultMessage: 'Title',
+  },
+});
 
 const blockRenderMap = Map({
   unstyled: {
@@ -17,6 +26,7 @@ const blockRenderMap = Map({
 
 const extendedBlockRenderMap = DefaultDraftBlockRenderMap.merge(blockRenderMap);
 
+@injectIntl
 /**
  * Edit title tile class.
  * @class Edit
@@ -31,9 +41,11 @@ export default class Edit extends Component {
   static propTypes = {
     properties: PropTypes.objectOf(PropTypes.any).isRequired,
     selected: PropTypes.bool.isRequired,
-    tile: PropTypes.string.isRequired,
+    intl: intlShape.isRequired,
     onChangeField: PropTypes.func.isRequired,
     onSelectTile: PropTypes.func.isRequired,
+    onDeleteTile: PropTypes.func.isRequired,
+    tile: PropTypes.string.isRequired,
   };
 
   /**
@@ -57,6 +69,26 @@ export default class Edit extends Component {
     }
 
     this.onChange = this.onChange.bind(this);
+  }
+
+  /**
+   * Component will receive props
+   * @method componentWillReceiveProps
+   * @param {Object} nextProps Next properties
+   * @returns {undefined}
+   */
+  componentWillReceiveProps(nextProps) {
+    if (
+      this.props.properties.title !== nextProps.properties.title &&
+      !this.props.selected
+    ) {
+      const contentState = stateFromHTML(nextProps.properties.title);
+      this.setState({
+        editorState: nextProps.properties.title
+          ? EditorState.createWithContent(contentState)
+          : EditorState.createEmpty(),
+      });
+    }
   }
 
   /**
@@ -85,13 +117,27 @@ export default class Edit extends Component {
     return (
       <div
         onClick={() => this.props.onSelectTile(this.props.tile)}
-        className={`tile${this.props.selected ? ' selected' : ''}`}
+        className={`tile title${this.props.selected ? ' selected' : ''}`}
       >
+        {this.props.selected && (
+          <div className="toolbar">
+            <Button.Group>
+              <Button
+                icon
+                basic
+                onClick={() => this.props.onDeleteTile(this.props.tile)}
+              >
+                <Icon name="trash" />
+              </Button>
+            </Button.Group>
+          </div>
+        )}
         <Editor
           onChange={this.onChange}
           editorState={this.state.editorState}
           blockRenderMap={extendedBlockRenderMap}
           handleReturn={() => true}
+          placeholder={this.props.intl.formatMessage(messages.title)}
           blockStyleFn={() => 'documentFirstHeading'}
         />
       </div>