+++ src/components/theme/View/SummaryView.jsx
@@ -7,7 +7,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Helmet from 'react-helmet';
 import { Link } from 'react-router';
-import { Image } from 'semantic-ui-react';
+import { Container, Image } from 'semantic-ui-react';
 import { FormattedMessage } from 'react-intl';
 
 /**
@@ -17,7 +17,7 @@ import { FormattedMessage } from 'react-intl';
  * @returns {string} Markup of the component.
  */
 const SummaryView = ({ content }) => (
-  <div className="view-wrapper">
+  <Container className="view-wrapper">
     <Helmet title={content.title} />
     <article id="content">
       <header>
@@ -52,7 +52,7 @@ const SummaryView = ({ content }) => (
         ))}
       </section>
     </article>
-  </div>
+  </Container>
 );
 
 /**